# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:9000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
    Bearer:
        type: apiKey
        name: Authorization
        in: header
paths:
...
  /change-password:
    # binds a127 app logic to a route
    x-swagger-router-controller: change_password
    post:
      description: Returns 'Change Password' to the caller
      # used as the method name of the controller
      operationId: changePassword
      security:
        - Bearer: []
      parameters:
        - name: "changePassword"
          in: "body"
          description: The changePassword of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/changePassword" 
          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /saveProjects:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_add
    post:
      description: Returns 'project_add' to the caller
      # used as the method name of the controller
      operationId: projectadd
      security:
        - Bearer: []
      parameters:
        - name: "body"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/projectadds" 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /FetchAllProjects:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_add
    get:
      description: Returns 'FetchAllProjects' to the caller
      # used as the method name of the controller
      operationId: FetchAllProjects
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /FetchFullProjectDatas:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_add
    get:
      description: Returns 'FetchFullProjectDatas' to the caller
      # used as the method name of the controller
      operationId: FetchFullProjectDatas
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /deleteProjectlist:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_add
    delete:
      description: Returns 'projects' to the caller
      # used as the method name of the controller
      operationId: deleteProjectlist
      security:
        - Bearer: []
      parameters:
        - name: projectid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fetchProjectByID:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_add
    get:
      description: Returns 'tasks' to the caller
      # used as the method name of the controller
      operationId: fetchProjectByID
      security:
        - Bearer: []
      parameters:
        - name: projectid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fullTimesheetProject:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetProject
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /fullTimesheetDate:
    # binds a127 app logic to a route
    x-swagger-router-controller: date_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetDate
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fullTimesheetAdminProject:
    # binds a127 app logic to a route
    x-swagger-router-controller: project_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetAdminProject
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /fullTimesheetAdminDate:
    # binds a127 app logic to a route
    x-swagger-router-controller: date_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetAdminDate
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  changePassword:
    type: "object"
    required:
      - "oldPassword"
      - "newPassword"
    properties:
      oldPassword:
        type: "string"
        example: "87654321"
      newPassword:
        type: "string"
        example: "12345678"
...
  projectadds:
    type: "object"
    required:
      - "project_name"
    properties:
      project_name:
        type: "string"
        example: "project_name"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string